Create README.md
This project is about searching and shopping for items on sky.com. It is built using Cucumber framework with core-java. The framework also use maven build tool with page-object architecture.
To run the test, navigate to RunnerCukeTest.java class and set the tags in the feature file for the right scenario, and the run the test. The has been designed to validate the test based on Junit assertion.